Move IpythonDisplay to top level utils folder
